아두이노 요청 (포트폴리오로 만들땐 삭제 필요) 1. /{id}/sensor-data url 기능 분리 단순 아두이노 센서 데이터만 받게
2.
period 아두이노 쪽에서 period 변경 요청

+++ src/main/java/com/example/plantServer/entity/Pot.java
@@ -25,13 +25,15 @@ public class Pot {
     private String serialId;
 
     @Column(name="humidity")
-    private Integer humidity;
+    private Float humidity;
+//    private Integer humidity;
 
     @Column(name="soil_humidity")
     private Integer soil_humidity;
 
     @Column(name="temper")
-    private Integer temper;
+    private Float temper;
+//    private Integer temper;
 
     @Column(name="water_level")
     private Integer waterLevel;
